pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"
[project]
name="cynthia"
version="0.0.0"
dependencies = [
    "about-time==4.2.1",
    "cynthia",
    "discord-py",
    "grapheme==0.6.0",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "pyotp",
    "roku"
]

[project.scripts]
cynthia="cynthia.__main__:main"

[tool.setuptools]
py-modules=["cynthia"]

